Release cand 1.1.3

+++ gtas-parent/gtas-webapp/src/main/webapp/admin/AdminController.js
@@ -13,13 +13,14 @@ app.controller('AdminCtrl', function ($scope, $mdDialog, $mdSidenav, gridOptions
   $scope.activeCodeTab = CARRIER; //sub-tab list under 'Code Editor' tab
   $scope.codeColList = {};
   $scope.isEdited = false;
+  $scope.OK = false;
 
   // CODE fields so we can generate the sidebar dynamically for each type.
   // could also pull these from the coldefs:
   // airportGrid.coldefs.map((col) => col.field), and filter the fields to exclude the "Edit" and "Id" fields?
-  $scope.codeColList.carrier = [{name: 'iata', len: 2}, {name: 'name', len: 255}];
-  $scope.codeColList.airport = [{name: 'iata', len: 3}, {name: 'icao', len: 4}, {name: 'name', len: 255}, {name: 'city', len: 255}, {name: 'country', len: 3}];
-  $scope.codeColList.country = [{name: 'iso2', len: 2}, {name: 'iso3', len: 3}, {name: 'name', len: 255}, {name: 'isoNumeric', len: 3}];
+  $scope.codeColList[CARRIER] = [{name: 'iata', len: 2}, {name: 'name', len: 255}];
+  $scope.codeColList[AIRPORT] = [{name: 'iata', len: 3}, {name: 'icao', len: 4}, {name: 'name', len: 255}, {name: 'city', len: 255}, {name: 'country', len: 3}];
+  $scope.codeColList[COUNTRY] = [{name: 'iso2', len: 2}, {name: 'iso3', len: 3}, {name: 'name', len: 255}, {name: 'isoNumeric', len: 3}];
 
   $scope.settingsInfo = settingsInfo.data;
   var that = this;
@@ -78,15 +79,13 @@ var setupCodeGrids = function(){
 
 var setCodeData = function (data) {
   var tab = $scope.activeCodeTab;
-  if (tab === AIRPORT){
-    $scope.airportGrid.data = data;
-  }
-  else if (tab === COUNTRY) {
-    $scope.countryGrid.data = data;
-  }
-  else if (tab === CARRIER) {
-    $scope.carrierGrid.data = data;
-  }
+  if (tab === AIRPORT) $scope.airportGrid.data = data;
+  
+  if (tab === COUNTRY) $scope.countryGrid.data = data;
+  
+  if (tab === CARRIER) $scope.carrierGrid.data = data;
+
+  $scope.OK = true;
 };
 
 $scope.refreshActiveCodeGrid = function(){
@@ -212,6 +211,8 @@ $scope.refreshActiveCodeGrid = function(){
           }
           else {
             $scope.refreshActiveCodeGrid();
+            that.successToast("Code successfully saved.");
+            $scope.OK = false;
           }
         });
       }
@@ -247,8 +248,50 @@ $scope.refreshActiveCodeGrid = function(){
 
     $scope.toastParent = $document[0].getElementById(tab+'Grid');
 
-    codeService.deleteCode(tab, $scope.rowSelected.id).then($scope.refreshActiveCodeGrid, $scope.errorToast);
-    refreshTooltips();
+    var r = confirm("Are you sure you want to delete this code?");
+    if (r == true) {
+      codeService.deleteCode(tab, $scope.rowSelected.id).then($scope.refreshActiveCodeGrid, $scope.errorToast);
+      refreshTooltips();
+
+      if($scope.OK) {
+        that.successToast("Code successfully deleted.");
+        $scope.OK = false;
+      }
+    }
+  };
+
+  $scope.restoreCode = function () {
+    var tab = $scope.activeCodeTab;
+
+    $scope.toastParent = $document[0].getElementById(tab+'Grid');
+
+    var r = confirm("Are you sure you want to restore this code to its original value?");
+    if (r == true) {
+      codeService.restoreCode(tab, $scope.rowSelected).then($scope.refreshActiveCodeGrid, $scope.errorToast);
+      refreshTooltips();
+
+      if($scope.OK) {
+        that.successToast("Code successfully restored.");
+        $scope.OK = false;
+      }
+    }
+  };
+
+  $scope.restoreAllCodes = function () {
+    var tab = $scope.activeCodeTab;
+
+    $scope.toastParent = $document[0].getElementById(tab+'Grid');
+
+    var r = confirm(`Are you sure you want to restore all ${tab.toUpperCase()} codes to their original values?`);
+    if (r == true) {
+      codeService.restoreAllCodes(tab).then($scope.refreshActiveCodeGrid, $scope.errorToast);
+      refreshTooltips();
+
+      if($scope.OK) {
+        that.successToast(`All ${tab} codes successfully restored.`);
+        $scope.OK = false;
+      }
+    }
   };
 
   $scope.createUser = function () { $location.path('/user/new'); };
